Renamed python files to prevent
expection after installation from package control related to import order
