+++ gist_20_exceptions.py
