+++ gist_60_helpers.py
