+++ gist_80.py
@@ -17,14 +17,14 @@ except ImportError:
 
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 
-from exceptions import MissingCredentialsException
-from gist_helpers import (
+from gist_20_exceptions import MissingCredentialsException
+from gist_60_helpers import (
     gistify_view,
     gists_filter,
     set_syntax,
     ungistify_view,
 )
-from gist_request import api_request
+from gist_40_request import api_request
 
 settings = None
